rename Modal to Dialog
