+++ packages/components/src/components/Dialog/Header.js
